pipeline/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
aksharamukha>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
